+++ lib/networking/default_response.dart
@@ -4,12 +4,17 @@ class DefaultResponse<T> {
     required this.data,
   });
 
-  late final T data;
+  late final T? data;
 
   DefaultResponse.fromJson(json, Map<Type, dynamic> decoders,
       {required Type type}) {
     assert(decoders.containsKey(type),
         'Your config/decoders.dart file does not contain a decoder for the following class: ${type.toString()} in modelDecoders');
-    data = decoders[type]!(json) as T;
+    dynamic jsonResponse = decoders[type]!(json);
+    if (jsonResponse == null) {
+      data = null;
+      return;
+    }
+    data = jsonResponse as T;
   }
 }