+++ lib/widgets/event_bus/update_state.dart
@@ -1,0 +1,13 @@
+import 'package:event_bus_plus/event_bus_plus.dart';
+
+/// UpdateState Model.
+
+class UpdateState extends AppEvent {
+  UpdateState({this.data, this.stateName});
+
+  final String? stateName;
+  final dynamic data;
+
+  @override
+  List<Object?> get props => [stateName, data];
+}